Add files via upload
Added some code which adds drivers of the current season to a list. This list is then used to modify an existing line of code which dropped all drivers with 0 points so that when rows are dropped from the df drivers of the current_drivers list are kept in the df even if they have 0 points.

Also cleaned up the code a little.

As of right now, I believe this code is complete (i.e. the actual game logic). I now need to get the game to a space that is playable for the user and make it look nice with a good UI.